requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
qrcode==7.4.2
mysqlclient==2.2.0
requests==2.31.0
python-dotenv==1.0.0
cinetpay==1.0.5
